requirements.txt
h5py==3.8.0
torch==2.3.0
pillow==10.4.0
tqdm==4.66.5
torchvision==0.18.0
scikit-learn==1.3.2
umap-learn==0.5.6
pandas==1.2.4
numpy==1.24.3